pva: mirror from gitlab cv/pva-sys-sw
Gitlab commit 0d38bbb765943 ("Revert "deploy: Update NvSCI p...")

Changes since last deployment:

0d38bbb7: Revert "deploy: Update NvSCI path in tegra tree"
2201a3ec: Revert "deploy: update nvidia-oot path"

Forked changes (dropped):

54b5ea83: Revert "deploy: Update NvSCI path in tegra tree"
1d42b346: Revert "deploy: update nvidia-oot path"
5819cf8b: Fix Misra defect of type 5.1
d9561abf: Fw: Fix Misra defects
ec8941c0: coverity: Qnx safety upload job timout increase
66fb1410: Fix pva_waiter_thread_func in umd tests.
0f06989c: Misra defect fixes in QNX KMD
76532814: Static defect fixes
1cbd5a36: Fix Misra defects
0deee89b: Fix misra defects
60335a28: Fix Misra defects
dbd5301d: Port pva_intf_test to pva_umd_tests
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/3508954
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>

+++ drivers/video/tegra/host/pva/src/include/pva_api_dma.h
@@ -82,8 +82,8 @@ struct pva_dma_transfer_attr {
 	/** When dynamic slot flag is set, it means the memory location will be
 	* relocated by commands.
 	*/
-#define PVA_DMA_DYNAMIC_SLOT ((uint16_t)1U << 15)
-#define PVA_DMA_STATIC_SLOT ((uint16_t)1U << 14)
+#define PVA_DMA_DYNAMIC_SLOT (1U << 15)
+#define PVA_DMA_STATIC_SLOT (1U << 14)
 #define PVA_DMA_SLOT_INVALID 0
 #define PVA_DMA_SLOT_ID_MASK 0xFFU
 #define PVA_DMA_MAX_NUM_SLOTS 256
